+++ src/main/java/com/github/aleksanderkot00/onlinesportsbetting/domain/dto/EventDto.java
@@ -2,6 +2,7 @@ package com.github.aleksanderkot00.onlinesportsbetting.domain.dto;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Data
@@ -11,6 +12,6 @@ public class EventDto {
     private String categoryName;
     private String teamOneName;
     private String teamTwoName;
-    private int teamOneScore;
-    private int teamTwoScore;
+    private BigDecimal teamOneScore;
+    private BigDecimal teamTwoScore;
 }
